Large number of changes.
#6 Some more fixes related to threads getting blocked, but probably not the ultimate fix. Did solve the problem for a few users though.

Added version number to logs and main screen.
Architecture changes, mainly for the input handlers to allow multiple options in the future. Root based options are in the works.
Added a voice based input handler that does not move around, but whispers the activated commands in your ear. Primarily for debugging.
Improved usability with a new interface that includes your IP address anda way to enable and disable the app.
Began framework for more options accessible via voice or touch.
Improved state management so we shouldn't have to rely on crashes much longer.